move test to another file due to conflicting fixtures
